Initial implementation of DW::Request, which is an abstraction layer around the Apache functionality.  This will, ideally, allow us to make the site a little more agnostic of what it runs on.
The basic idea is that you implement the logic in DW::Request about how to
know which module to load, and then you implement the module as something
in that DW::Request::XXX namespace.  Look at DW::Request::Apache2 for an
example.

The site needs to be converted to use this module, but it doesn't have to
happen immediately.  Both methods (using the Apache modules directly and
using the abstraction layer) will live in perfect harmony for now.

Look at the ljlib.pl changes for examples of using this module.