http://bugs.dwscoalition.org/show_bug.cgi?id=80
Initial commit of WTF backend.  This is expected to destabilize the base of the
site as there will be a lot of things that don't work right now.  You should
expect everything that adds/removes "friends" is now broken until it gets
ported over to use the appropriate new methods.

The nomenclature is fairly straightforward: watch and trust.  Only people and
identity accounts are allowed to Watch, and only people accounts are allowed
to Trust.

You may Watch any account type and entries made in those accounts will show
up on the page that shows aggregated content (your reading page).

You may Trust only people and identity accounts, which gives them access to
view content you have marked as "Trusted Only".

Presently, the "manage_circle" console command can be used to add and remove
people from your watch/trust lists.  If you update your watch list, content
will start appearing on the "/friends" page.

If you trust people, they should have access to your protected content.  This
is not warranted to work in all cases, and a production site should not be
deployed on this code yet.  There is still a ways to go with testing and UI
functionality, but at least the core is in.

Patch by Mark Smith <mark@dreamwidth.org>.