http://bugs.dwscoalition.org/show_bug.cgi?id=371
Completely update trustmask handling logic in WTF system to be better documented code and properly handle not overwriting when the user is not changing the trustmask.  Also fix test-wtf which broke after the recent change to remove_edge.  Also resolved issue with tester with memcache turned on.

Patch by mark.