http://bugs.dwscoalition.org/show_bug.cgi?id=2054
Add support for a routing system and Template Toolkit.  This is Apache2 specific right now but as that's all we support, that's fine.  Nothing yet uses TT, either, but the edges data page is now using the routing system.  This should be considered experimental while Andrea (mostly) and I (some) work on it.

Patch by exor674.