http://bugs.dwscoalition.org/show_bug.cgi?id=2080
This is a logic change/update for accessing (viewing, editing, and deleting
posts).  This will mostly impact OpenID users as this will give them a lot
more access than they had before.

Notably: when we do a community import, the OpenID accounts that now own those
posts will be able to manage the comments on them, as well as edit and delete
the posts in question.

This should have no impact on regular users, but as always with a change to
such core logic, there may be some unintended consequences.  Please test.

Patch by mark.