http://bugs.dwscoalition.org/show_bug.cgi?id=2544
If the community allows everyone (even nonmembers) to post, then show the post link in the navstrip.

Also syncs the translation strings in the code with the ones that are live on the site (which are shorter), to prevent the navstrip from stretching the page too much.

Patch by denise.