http://bugs.dwscoalition.org/show_bug.cgi?id=2739
We used to pass in db handles; then it shifted to maybe we were passing in db handles (or maybe not). And now we don't pass in db handles at all -- so remove the code which used to take care of the maybe case. It's no longer needed.

Patch by kareila.