http://bugs.dwscoalition.org/show_bug.cgi?id=930
Improve tag cloud HTML: use a ul/li, not spans. Add CSS so that the list items display inline. Add a rel="tag" to the tag link. Add an invisible element with the number of uses so the information is exposed to screen readers, which won't be placing any importance on the font size.

Patch by deborah.