http://bugs.dwscoalition.org/show_bug.cgi?id=2317
Two token swap of usernames; new page linked to from /rename, changes to community renames so that you can only rename a community (or rename to a community, because it then gets renamed to ex_*), when you are the only community member. This is to avoid surprises.

Patch by fu.