http://bugs.dwscoalition.org/show_bug.cgi?id=3687
Replace references to ljconfig.pl with etc/config.pl. Fix a bug in startup where we don't properly set the modification time of the config files to be checked in case the files need to be reloaded later (we only checked the config.pl when we should really be checking all config files)

Patch by kareila.