Allow local package installation
This is mostly for my development environment. It allows the usage
of cpanm (http://cpanmin.us/) to install packages locally as the dw
user. These packages are installed in $LJHOME/extlib/ by convention.

I'm testing this out and seeing if it's a good plan for production or
not. It might be interesting because then we can just push the extlib
directory out, possibly? I'll experiment with it.

Patch by mark.