http://bugs.dwscoalition.org/show_bug.cgi?id=4073
Adds a filter by is_deleted, to indicate when the journal is not visible. This bit is changed when the user deletes/undeletes or is suspended/unsuspended. It's only recognized after the indexer has been rerun, though, so also add an additional check to the output in /search.bml (which is uglier, but reveals less info). Then, also make sure we delete from posts_raw when the user is purged -- the entries don't show up anyway because the user was previously deleted, but this cleans up.

Patch by fu.