http://bugs.dwscoalition.org/show_bug.cgi?id=2524
Make the Create Poll link context-sensitive: show it only when your journal caps let you post a poll, or when you're explicitly using a community that has the post poll capability. (Seems misleading to free users otherwise, since it's not clear that they *might* be able to post a poll to a paid community they're a member of)

Patch by fu.