http://bugs.dwscoalition.org/show_bug.cgi?id=1044
Shows a warning and asks you to submit again to confirm that you really want to do this, if user is expunged/purged/suspended., and you're trying to buy  paid time for them. And bonus: remember the last form field values when there's an error, because it's annoying to ask you to resubmit otherwise.

Patch by fu.