Fix race condition in the payment system
There's a race condition here that can sometimes cause us to mark a
payment as having an internal error when in reality it is just in that
magic moment after Gearman has finished processing it but we haven't
registered the change in the database yet.

This fix adds a second check to check the database again to see if it's
been updated. Also, I have to say, this fix was suggested by Denise, I
just implemented it.

Patch by mark.