http://bugs.dwscoalition.org/show_bug.cgi?id=4201
Initial support for claiming an OpenID account.

The only shortcoming of this, right now, is that it doesn't update
talkleft. This means that the "you've posted 2349822 comments" text will
be wrong on the profile pages. I'm not sure how I want to fix that yet.

The actual content will be fine, though.

Patch by mark.