Updates to Comment importer
This adds XML caching -- it's not great and it's not done, but it works OK and I need to get it out of my tree. Biggest problem right now is that it never removes the cache files so if they go stale, the import will "succeed" but won't get new stuff.

Patch by mark.