http://bugs.dwscoalition.org/show_bug.cgi?id=4177
Use camel case for the function name; in inline onclick handlers, the variable name randomicon was being treated as the element id, not as  a function reference, so we want the  function name to not match.  Changes to input type='button'... rather than a button so we don't submit the form. Also remove extra whitespace which was causing the buttons to look unbalanced.

Patch by fu.