http://bugs.dwscoalition.org/show_bug.cgi?id=4249
Merge the list of props to be preloaded, and the list of props to be saved. This makes both lists fully consistent, and reduces duplication (less chance for errors to be introduced). Added: opt_whatemailshow which was only in the "save" list before. Removed country, state, city, sidx_loc, timezone from the preload list, because these are handled by the Location widget.

Patch by fu.