Don't backdate community posts when importing
We use the backdate option so that posts you import don't flood your
reading page. It turns out that this is unnecessary for community posts
because we manually set the logtime/rlogtime to an old value. This means
we can turn off backdate and we won't flood the reading pages.

Patch by mark.