Markdown ninja-commit (testing)
Discussion and hack time at YAPC::NA has caused me to implement this
(hacky) way of using Markdown on Dreamwidth. In short, if you make
the first line of your post be "!markdown", the rest of the post is
interpreted by the Text::Markdown module.

This also enables a version of the testing user-reference code that I
am playing with. This is the Twitter style of "@foo" referring to a
user. You can also do "@foo.livejournal.com" to refer to a LiveJournal
user.

Most user tags that you would originally write like this:

    <user name="foo" site="foo.com" />

Can, if Markdown is enabled, be this:

    @foo.foo.com

NOTE: this is being committed to give people the chance to play with it
if they want. The implementation is subject to change after discussion
and testing.

MORE NOTE: this only applies to entries. We could add this to comments,
but haven't yet.