Respect minimum security on email posts/images
There were old emailpost_img* userprops. They don't exist anywhere else in the
code outside of these few files, so they weren't available for users to set
or even respected.

They existed separately because of FotoBilder's separate stack. It didn't have
the same security system as LiveJournal. Since Dreamwidth is building photo
hosting in to the core, I've collapsed this.

Security for images posted by email is now the same as the post that contains
the images. There is no option to have public photos with a private post,
for example.

Further, post-by-email security now works like this:

    * If you specify a security tag in the message, we use that.

    * If not, we will use a specified "post-by-email security". This is the
      emailpost_security userprop which is set on /manage/emailpost.

    * If not, we will use your account's newpost_minsecurity userprop, which
      is the one most people are familiar with.

    * If none of the above are present, we default to public.

This should meet with people's expectations. If not, please let me know.