(Bug 4570) Inactivate userpics
In order to import everything we can, right now we allow the importer to
exceed the user's userpic quota. This adds a nightly update that goes in and
makes sure the userpic quotas are respected.

This runs as part of the nightly active user job, so it means users will have
up to 24 hours of time above the quota before we fix it.