(Bug 2524) Only remember value if user did the selecting
Handles this case:

- Set your journal to a minsecurity of "Private"
- Go to make a new post
- Pick any community that has a lesser minsecurity (anyone, members)
- Security of post will remain "Private" instead of that comm's
  minsecurity

This was better than the alternative of, if you'd selected "Private"
manually, it being switched to a lesser security, but honestly neither
was ideal.

So now instead of looking at the current value before adjusting things
for minsecurity, it becomes smarter and records a value when the user
themself has selected an option, and uses *that* value (only) if it
exists.