(Bug 4679) Don't mark Recent Entries when browsing tags
When browsing entries in a journal by tags, the "Recent Entries"
part of the navigation menu was shown as the currently active one
by appliying the "current" CSS class. When browsing tags, the current
class should bot be applied to any of the menu items.

This tags sets the "view" attribute of the S2 page object to an empty
string if the user is browsing by tags. Thus the "current" class is not
applied to the "Recent Entries" div by the function
print_module_navlinks.