(Bug 4660) Add explanation to username swap error
Adds a bit more explanation to the error that occurs when a username
swap fails because the two accounts do not have the same verified email
address and the same password.

The community error does not need to be changed because with communities
"under your control" is clearer than with personal jounals - it means
being the community maintainer.