(bug 4720) Changes error message when a search tab has been open too long.
If you leave a search tab open for more than 24 hours, check_form_auth
fails. This changes the error message to be a little clearer and tell
the user how they can get it to work again. Patch is UNTESTED because I
don't have search configured on my 'hack, but it only changes the text
so I don't think it's too dangerous. :)