(Bug 4780) Update the tracking JS to use dialog + new ajaxtip API
Also fixes the issue where the comment icon didn't update when
tracking a thread.

It was broken because we passed in the LJ_cmtinfo variable when creating
widgets (so we could access the cmtinfo as a local variable for slight
efficiency purposes). Problem was that the LJ_cmtinfo variable is
initialized after we create the widgets. So now we just use the global
variable, which isn't ideal but will do.