(Bug 2174) Reading filter options: public/order
I went in to do Bug4501 and found that Mark's excellent set up already
had this support in the backend, where it counted.  Public filters
and sortorder were already supported and just needed to be added
to the UI.  I know we want to overhaul this UI eventually, but in
the meantime, users will be able to set these filters as public
and order them how they wish.  This additionally satisfies
Bug4501--in the interface, the filters in the drop down are
sorted by sortorder first, and then name. Someone who hasn't set
any sortorder for any filters will see them ordered by name.