Revamp importer logic to drop URLs
We've once again been bitten by the bug wherein we depend on the URLs
and end up causing duplicate posts or comment imports. While it was nice
to depend on those URLs, we don't need to.

This commit changes it so that our import_source lines are actually just
slash separated tuples that we can generate with information that we
know, and that the remote site has to provide. (Itemids, notably.)

This commit also takes care of a long-standing issue and updates the
grabbed_until columns for jobs so that long running jobs won't get
passed on to a new worker while they're still busy.