(Bug 4708) Hide the first set of buttons, not the second
The first set of buttons are duplicates which only show if we have
filters for the account. The second set of buttons always show. What we
want is to have the first set of buttons show if there are filters, and
in addition, if the subscribe/etc are selected (that is, the filters
could be visible)

Just changes which set of buttons are affected by the JS, logic
otherwise remains the same.