(Bug 1960) Test LJ::Tags:is_valid_tagstring
Easier than manually testing since we don't have to reload the page and
try to remember regression testing on futher changes.

Some basic tests, and one specific to this bug, "truncated right at a
trailing space", which confirms that we can no longer save a tag with a
trailing space, even if the trailing space only is there after we've
truncated.