(Bug 4719) Coerce all false values to the numerical 0
We can't rely on defined-ness. The cached value is stored as an empty
string when there are no points, so it fails the defined test and
doesn't get set to 0. Then in the frontend it gets printed as an empty
string...

This means that on production, or anywhere else with memcache on, first
time you load the page you'll get "you have 0 points". On subsequent
loads it becomes "you have points". Misleading! Annoying! Weird.