(Bug 4855) Get the first non-DOOO entry when importing
Getting the latest entry via the protocol gives us the latest entry
according to eventtime (user-provided time) rather than logtime.
Sometimse this could be a DOOO entry, which then means that everything
after that DOOO entry will not be imported. So let's try to find the
first non-DOOO entry, as that will be the latest according to logtime.