Add comment searching/indexing
This patch updates our Sphinx setup so that it indexes and searches
comments. They are intermixed with the normal results, sorted by the
sort order the user requested.

This patch also changes the way we index content so that when a post
or comment is posted/updated, we only insert that particular item into
Sphinx instead of doing a full sync. This should keep the load low even
on the busiest communities.

OTOH, this means that we don't automatically fix up accounts that got
badly indexed for some reason. We will need to give support admins a
tool to kick off a full reindex of an account.