(Bug 4934) Abort the ajax request if user moves away from the trigger
If the user moves away from the trigger while we're still loading, the
loading symbol disappears, but then a few moments later the popup with
content shows up. In this case, we should assume that they're no longer
interested, or triggered it by accident, and prevent the request from
going through.