(Bug 2152) Made changes to Entry.pm following on from code review discussion.
This involved alterations to module-sticky.tt, most notably to change the value of "sticky_select" form elements
so that their results could be cast into an integer describing the desired position of the sticky (O indicating the
entry is not a sticky).  This made the use of the form element more intuitive in Entry.pm.  Previously sticky
position had to be determined by comparing the ditemid of an edited entry with that of sticky entries, not it
can be determined from the form variable.  Variables in Entry.pm were re-named accordingly where appropriate.