(Bug 5025) Enables markdown parsing for replies-by-email
So. Gmail automatically hardwraps sent messages. We can strive to unwrap
it, with the attendant difficulties (what if the wrapping is in the
middle of an HTML tag? what about multiple line breaks vs single line
breaks?), or we can rely on a library which is meant to handle text like
this.

So. Markdown -- HTML still works, along with the various formatting
shortcuts markdown offers.

For a certain subset of users who regularly use their mobile phones,
this should also be easier than switching keyboards to type angled
brackets/quotes

We use regular markdown, with a variant of @username to refer to a
specific user.