(bug 4772) rework code that creates <li>s
Each element of @$accounts is a hashref like so:
 { userid => 1234 , timeupdate => '2013-03-30 01:58:32' }

Then LJ::load_userids gives us a hash that lets us map userid numbers
to user objects.  Unfortunately, we need to keep the @$accounts array
around to preserve the ordering -- they're sorted by last update time.

Hopefully subbing in the conventional $u will make what is going
on here clearer.  If not, we'll have to rework this in a more
readable but less efficient way. ^_^;

Addresses afuna's comment:
https://github.com/dreamwidth/dw-free/pull/337#discussion_r3792930