Change MEDIUMINT to INT for comments/counters
Note: this does not include the normal "upgrade" pathway, because
actually upgrading these columns is a very time-consuming operation and
not needed by almost anybody. If someone needs to expand their columns,
they can do it manually.

The reason for changing the table creation commands, though, is to
future-proof further Dreamwidth clusters, and to make it so future
installations use the wider column widths.

There are no code implications by having the clusters
widened. Dreamwidth runs with one cluster that uses MEDIUMINTs and one
that uses the INTs. The code doesn't care anymore (it used to, but that
was fixed a while ago).