http://bugs.dwscoalition.org/show_bug.cgi?id=4927
Added the "More Options" button to the comment preview page with jQuery
as well as making the functionality. Also edited Talk.pm to resolve a
bug in my original code that made this break unless you came to the
preview page via QuickReply.

Additional stuff: removed a sexist joke from the comments in S2.pm and
added "fixme" comments to viewing_style_args and viewing_style_opts in
Web.pm as these need to be modernised (as advised by Rah).