(Bug 4715) Media upload form
Image upload form.

This also merges the generate id and upload step.

This adds the code that parses image uploads via multipart attachment.
This also wires the controller to the backend, which allows the frontend
to actually upload files!

There's also one sneaky bit of property related code that does nothing
so far.

Also, removes a convenience method in DW::Request which overrode the
official getter/setter.

Also make sure that we return a JSON error message if not logged in.

This submits the metadata via JS:

* submit when the user says to submit
* resubmit, if needed, after all uploads are finished
* tweak styling of the upload form slightly -- don't hardcode colors
* use classes instead of an id