(Bug 215) Fix up code for vgift purchase in contextual hover.
This provides a valid value in the ctxpopup endpoint for the existing
can_receive_vgifts check using our can_receive_vgifts_from user method
(which won't ever be true unless vgifts are enabled in the site config,
so it can safely be used in production ahead of deployment).

The sendvgift.href (and jQuery equivalent) are updated to use the proposed
new shop URL format of /shop/vgift?user=username, instead of vgift?to=username
that LJ apparently used.  This logic is encapsulated in the new user method,
$u->virtual_gift_url, and encoded in the endpoint as url_vgift.

Small logic tweaks in the can_receive_vgifts_from user method ensure that
the logged in user is a personal account, and the proposed recepient is not
a syndicated account.  Also added explicit check for journal bans.  This was
already being done in the shop, but in testing I realized it would still show
the option in the contextual hover menu for a banned user.

I thought about also checking for the case where the logged-in user
has banned the other user, since seeing "Send a virtual gift" right
above "Unban user" in the menu seemed odd.  On further reflection, I
realized moving "Send a virtual gift" up right below "Send message"
would look less weird.  But once that was done, the link begged for
more consistent grammar so I changed the message to "Send virtual gift",
removing the indefinite article.