(Bug 215) Add database table for vgift transactions.
This is the last remaining table from the initial database design, which was
drafted in 2009 but not needed until now.  Some properties of the transactions
are stored in the purchase cart instead of the transaction record; I didn't
want to duplicate information, but we may want to revisit this later.

Because this is a clustered table, it is added to the list in @LJ::USER_TABLES.
There is already a key for vgifts in alloc_user_counter, so we just change it
to use the correct columns from the new table.