(Bug 215) DW::VirtualGiftTransaction - methods for working with transaction data.
This was originally implemented in the context of DW::VirtualGift, but
as the project evolved, it was clear these would better be handled as
methods on a new kind of object representing a single vgift transaction.

Several rounds of tweaking have been incorporated into this one patch.
If anything described seems to be missing, it should show up later.
The FIXME lines in particular will go away with the addition of another
new module defining a new shop item class.