(Bug 4866) Allow collapsing of all profile sections to stick
This changes profile_collapsed_headers in userproplist from
a char to a blobchar - char is limited to 255 characteters, and
with the number of sections we have, not all would fit and the
last property would get truncated, which means existing data
is potentially corrupt.

As such, this change does not migrate old data, and users will
have to reset the collapsed sections after it goes live.