(Bug 5265) Move greeting/footer logic into LJ::format_mail
Much easier to test, because you know the full text of the email (rather
than having to guess if you need to put a newline or not).

For the record: no need for additional newlines before/after the body.

Patch includes whitespace fixes in the text of the test emails