Explicitly detect `isremote` for the journalselect event
(And do so only when we're "posting as other")

We'd assumed that a change on #postas_usejournal or #post_username would
only be triggered if a user selected "post as other", therefore isremote
would automatically be false.

However, Chrome started firing a change event on #post_username even
without user input (perhaps to do with autocomplete?). So now we
explicitly check the current remote against the username of the user
we're posting as..

But we can't just compare postas_usejournal to postas_remote because
they could be different because of autocomplete. So we also make sure
that all this only happens when the "post as other" radio box is
selected

Fixes 709.