(Issue #821) Respect edge blacklist, lower limit
Data obtained by /data/edges can be quite large; limit the data to 5000
edges for most edge types. Also, respect the
%LJ::FORCE_EMPTY_SUBSCRIPTIONS configuration option.

This patch also fixes a potential race condition and adds functionality
to the member_userids and member_of_userids functions to request a limit
to the number of userids returned.