Implement Previous/Next Entry links within tag. Fixes #943
Building off of the existing code base for the generic previous/next
entry links.  The new links use the same structure, but pass in one
additional argument (the tag to search within) to go.bml. If a tag is
passed in, it narrows the SQL query in get_item_id_near2()
appropriately.