[#665] Remove extra bottom padding take 2
Previous solution led to an annoying jiggle on page load. Also, hacky!
So let's do it this way: use jQuery to detect the last visible row and
set the .last-visible class for a component that has extra elements that
are hidden/displayed by JS. If not, then the original CSS which uses
:last-child works just fine

(If there's no JS, we only have a little extra padding for specific
components -- not a dealbreaker)