Remove blockwatch and docs
We don't use it. It's another profiling system that we haven't kept up
with and is rotted. If we ever get crazy huge, we can reinstate it and
bring it up to speed.

This also removes most of doc/raw/ which was a bunch of crufty old
documentation we don't maintain and barely applies to Dreamwidth these
days. Please refer to our wiki for documentation.