Make cleaner closer more intelligently
There is a problem where the HTML cleaner just randomly closes tags in
some ineffective order. Instead, we should actually use the tag stack to
close tags in the order we saw them.

This also removes autoclose as an option: we always want to close when
we're doing HTML cleaning.

Similarly, this removes tablecheck: we always want that on. If someone
is throwing in table related HTML, we should ensure that it can't get
busted out somehow.