[#1340] default config does not contain refund rate for paid accounts
I had trouble testing the previous commit because the config values for _refund_points
were never committed anywhere. Without that information, it prints an unhelpful error
message: "Sorry, your account type is not eligible for a points conversion."

This adds the values we set in production: 30 for paid, and 41 for premium paid.