[#1315] allow spamreports to handle IPv6 addresses
This patch does two things:

1. Widens the column for IP addresses in the spamreports table from 15 characters to 39.
An IPv6 address may contain 8 16-bit colon-separated hexadecimal codes for a maximum
width of 8*5-1 = 39 characters.

2. Allows the spamreport page to load for any given IP value that is 39 characters or less.
If the IP does not exist in the spamreports database, the "noreports" error will be shown.
If the IP exists, the report may be viewed and closed, but the sysban option will not be
presented unless it is a valid IPv4 address.  Once we verify sysban support for IPv6,
this clause may be removed.

Fixes #1315.