Preserve sticky-entry value when module unloaded
Before, there was no way to distinguish between "user left checkbox
unticked" and "checkbox doesn't exist" for the sticky-entry checkbox,
resulting in formerly-sticky entries losing their status.  This fix
wraps the remove logic in another if statement, so that it will only
perform the remove if the checkbox existed to be unticked.

Fixes #1343