[#1315]	tweaks based on	alierak's comments
Removed	the proposed verify_ipv6 subroutine, since it doesn't capture all edge
cases and isn't directly relevant to the main purpose of the change, which is
to allow non-IPv4 spamreports to be closed.  Also further widened the IP column
of the spamreports table from 39 to 45 characters, just in case we ever receive
any hybrid IPv6/IPv4 address formats.