Fix alignment issues in the entry actions
Some irritating alignment issues; should be fixed in relatively modern
browsers. Older browsers use the original (now fallback) code, which
means that in some cases there's some misalignment, but not enough to
impact usability.

Note: it would have been nice to be able to set width: 100% on the
dropdowns, but in webkit browsers, this leads to the items combined
taking up 100% of the width each (so e.g., post button + dropdown going
off the edge of the page)