Add seperate header link color variables to all layouts
Adds variables for header link colors to Practicality and Bases
and adds them to existing themes, using the same colors that were
being used.

Note: this will not set custom colors if users are using them,
and they will have to set the new variables themselves in the wizard.

Fixes #677