Rescreen comment edits
For entries which screen comments, editing a comment which had been
unscreened would immediately display the edited comment. This patch
changes the behavior such that an edited comment is immediately
rescreened, for the post owner(s) to review the changed content and
determine whether it is still eligible for unscreening.  The user
will be warned that their edited comment is being rescreened.

Fixes #1230