fix t/formatted-mail.t by tweaking behavior of DW::External::Site::get_site
This was a tricky one to track down.  The git bisect tool told me to look at
the changes to DW::External::Site::get_site, but it was not immediately
obvious why this was the culprit.

It turns out that because the name of the site in the test config is
"test.dw" this was getting mapped to site="dw" and generating
dreamwidth.org user tags instead of test.dw user tags.

Obviously .dw isn't a valid TLD, but a better approach would be to
assume that if there are no dots in the site argument, it should
be mapped directly to the site alias table insted of being treated
like an internet domain name.