[#1468] let LJ::OpenID always use the paranoid useragent
This removes the IS_DEV_SERVER special casing that resulted in the
paranoid useragent only being used in production for some functions
of LJ::OpenID.

The bypass for the blocked_hosts check is left intact as a comment,
in case it is needed for testing purposes at a later date.

Fixes #1468.