[#1483] explicitly specify UTF-8 charset
If a developer's LANG environment variable uses something other than
a version of UTF-8, compiling SCSS files that contain Unicode characters
will fail with invalid character errors.  The easiest and best way to
fix this is to explicitly specify the expected UTF-8 charset on the
first line of each affected file.

Fixes #1483.