Hide subscription options for non-visible accounts
When offering tracking subscription options for an account with a status
other than "visible", only offer notification of when the account is
purged.  In particular, don't offer notification of when a new post with
a specific tag is made, since that includes the list of tags from a
deleted account.

Fixes #1523.