[#1356] remove "identity_posting" userprop and related checks
The legacy userprop "identity_posting" was intended to allow community
admins to opt into allowing OpenID users to post entries and manage
tags.  Since we decided to allow this for all OpenID users across
the site, this removes the userprop and all the permission checks
that were failing as a result of it not being set.

The relevant tests in protocol.t have been updated, but that test
suite is still set to be skipped because of inconsistent results
in other parts of the test suite.

Fixes #1356.