[#1628] check for userids, not IP addresses
This changes the conditional for deciding whether an antispam report
is for a user or an IP address.  The old regexp only looked for
the existence of a period, which worked for IPv4 but not IPv6.

Since userids are only digits, use a regexp that checks for those
instead and assume anything else should be treated as an IP address.

Fixes #1628.