Add front-end for FAQ version control
Add a "Version control" section to the bottom of edit page for existing
and deleted FAQ entries, which lists entries from the faqhist table by
date and editing user.  When selecting a revision, load the normal FAQ
editing page, but populate it from the faqhist table rather than the faq
table.

Fixes #1624, which is blocking the merge of #1516 to fix #1455.