[#1689] add filter selection menu to mobile read page
We can't use the "/read/groupname" format URLs since this is
a BML file in htdocs and is restricted to using getargs, but
I doubt most mobile users want to hand-edit the URL anyway.

The old LJ behavior was to try to show a filter named "Mobile
View", then "Default View", then all items.  This adds a
selection mechanism modeled on the dropdown menu from the
control strip, which also includes journaltype filtering.

Anything fancier can wait for the day when we either redo
the entire page or scrap it altogether.

Fixes #1689.