On tag-edit failure, warn but allow rest of edit.
Previously it would present an error despite having applied *some of*
the rest of the edit, e.g. body but not props/slug.

The change to Tags.pm means if a tag removal is rejected, any tag
additions also don't happen. In principle you could leave out that
change, but it didn't look trivial to make edittags.bml behave
consistently without the change (if you're going to mostly-succeed you
should still end up at the entry page, which we can't easily put a
warning at the top of).

Fixes: #1528