remove backend support for unique-email polls
D has expressed surprise at least once upon discovering that the backend code
for unique-email polls still exists. It was designed to prevent multiple users
with the same email address from voting in the same poll: "It was only ever used
for some LJ 'official' polls like voting for the advisory board."

If we don't want the code, we should remove it.