Move the inbox's expando activation event to the link
The link is what makes the expando button selectable from the keyboard,
but the event was on the img tag.  That works fine for clicking, but
results in keyboard-activation following the link instead of firing the
desired event.

This is a partial fix for bug #1808.  It fixes keyboard activation if
JS is enabled, but does not attempt to deal with the no-JS case.