Correctly render special characters in PM subjects
Fixes #1652.

Depends on #1811 -- this makes use of the LJ::dhtml function introduced
in #1811 to ensure that special characters in PM subject lines are
preserved on reply, instead of ending up in an endlessly-nested sequence
of &amp;amp;amp;amp;amp;..., making use of the fact that special
characters do not need to be encoded as HTML entities in the compose
form.