Start BlobStore
This starts implementation of BlobStore which will replace all the
various storage layers we use for userpics/media. The idea is that we
should have one simple storage system and offload the work to other
people instead of doing it all ourself.

This is first part in removing MogileFS. It's a really complicated
system that is better replaced by using something simple with frontend
caching and/or a file storage service other people provide.

(Although the BlobStore interface was made with the idea that it can
support MogileFS pretty easily, so it's not actually hard... if we added
'namespace' to the API then it'd be a drop-in replacement.)

Additionally this adds Log::Log4perl as an easy way of logging things in
a more uniform way. I'm tired of not having logging messages and this
adds a very basic method of logging (with Carp support) in such a way
that it shows up in the Apache logs, but in a very useful format that
includes the package/line number that is generating the logs.