Per-user Selective Screening (#1832)
New methodology to allow forced screening of all comments from
a particular user. In terms of backend, functions much like a ban, but
with neutral wording to reflect the fact that this behavior is desired
for many users, and not a negative consequence. There is not (yet) an
equivalent page to /manage/banusers for this functionality, so for the
moment it is only accessible through the admin console.

Fixes #1234