[#1848] verify that 0/zero is a valid textcaptcha answer (#1880)
I made a copy of the test for multiple answers that shows
that our DW::Captcha logic for zeroes should be fine, even
if the zero passed in is a number and not a string.  If we
are seeing an issue, we can be pretty sure that the problem
isn't in our code.  We're ultimately testing the value of
an MD5 digest of the answer, not the answer itself.

Fixes #1848.