[#1896] only populate notifyarchive if esn_archive enabled
Metadata for every notification is saved in this table,
but if the feature is disabled, it's never used anywhere.

Perhaps this "feature" could be entirely removed in the
future, but I think that decision should be delayed until
we have a spec for the requested inbox redesign.  It might
be useful when trying to implement a trash folder, for example.

Fixes #1896.