[#1928] use eval to catch croaks from DW::Media->new
In get_active_for_user, the process of loading the
user's images would die if DW::Media->new failed
for any of the images.

This allows the process to complete and the page
to display normally minus the problematic images,
trapping any image errors in an eval statement
and printing them to the error log instead.