[#1928] create image in MogileFS before writing to db
This changes the order of operations in upload_media
to try to create the MogileFS file first - that way
if we get the "unable to instantiate file" error,
we haven't already created bogus rows in the media
and media_versions tables that would cause problems.

In order to do the MogileFS operation first, we need
to use the mogkey object method to get the file key,
but we can't use the new object method before the db
rows exist.  The workaround is to bless a fake object
into the class using the minimal information needed to
use the mogkey method.

Yay for reducing the number of FIXME comments.