add compile checking of command line scripts
In the past, we've only done compile checking of modules
in cgi-bin.  This adds similar checking for scripts in
the bin directory.

Since we don't actually want to execute the scripts,
we change the test condition to execute 'perl -c'
instead of 'perl' (which was OK for modules).
This only affects files ending in .pl -
files ending in .pm are tested differently.