Add S3 support to BlobStore
This implements S3 support for BlobStore using the Paws SDK, which seems
to work fairly well. This also updates the BlobStore initialization code
to not be magical about what fallbacks are valid, it's now just an array
that can be configured directly by the admin.

AWS credentials are by default handled using IAM roles/instance
profiles, which is the secure way to do it if you are running a server
in EC2. This is highly recommended, but fallback support is provided for
specifing the keys directly (for now?).