fix undefined warning in DW/Controller/ChangeEmail.pm
Use of uninitialized value in string ne at /home/dw/production/cgi-bin/DW/Controller/ChangeEmail.pm line 84.

First, we don't need to check is_identity when assigning $password,
because we check for it in the conditional.

Second, we should actually use $password if we assign it.

Third, put the short-circuit check ($remote->is_identity)
first, to avoid unnecessarily checking a string that will
be undefined in that particular case.

Fourth, OK to error if $password is undef, since that can never be
the correct password.  Identity accounts can have undefined pws
but we've already accounted for that.