use another method to repopulate the entry object data
At first I thought I would need to audit the code for
uses of $entry->{security} instead of $entry->security,
and while I did find and change a few of those, I quickly
realized a less scattershot approach would be to drop in
another method call in place of the removed one that had
the same side effect of populating the data in the cached
entry object.

With that done, the failing test once again passes
(although I'm still committing the fixed version).