additional headers for media file requests
Discussion with @alierak led to the conclusion that media files
should use the Cache-Control and Last-Modified headers.

Inspection of Apache/LiveJournal.pm and Apache/BML.pm indicates
that Cache-Control should be applied to all user content responses,
so we make that the new default behavior for user content in DW/Routing.pm.

The Last-Modified header is set in the media file controller as the
upload date of the file, which will never change.

Without the Last-Modified header, the image may be redownloaded
unnecessarily.  Without the Cache-Control header, Cloudflare
caching may interfere with the ability to load the image correctly.