[#2003] also use error header for Cache-Control
@alierak explains that there are two header tables, one
for successes (which header_out uses) and one for errors
and redirects (which err_header_out uses).  So we need to
make sure that user cookie bounces use err_header_out to
avoid caching.

He says we could probably set this header on every request,
but let's start with the one place where we know we need it.

(Most of this is whitespace cleanup; the last line is the
bit that does the thing.)